avance1

+++ src/main/java/com/example/ProyectoBancoJPA/repository/TransaccionRepository.java
@@ -3,5 +3,5 @@ package com.example.ProyectoBancoJPA.repository;
 import com.example.ProyectoBancoJPA.model.Transaccion;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TransaccionRepository extends CrudRepository<Transaccion,Integer> {
+public interface TransaccionRepository extends CrudRepository<Transaccion, Long> {
 }
